+++ packages/svg-composition/rust/src/core/modules/svg_render/resources/svg_composition/svg_element/attributes.rs
@@ -16,7 +16,7 @@ pub enum SVGAttribute {
         height: u32,
     },
     Opacity {
-        opacity: u8,
+        opacity: f32,
     },
     Transform {
         transform: SVGTransformAttribute,
